Create Meter Page

+++ MntPlus/ViewModels/Controls/ReadingMeterViewModel.cs
@@ -30,20 +30,67 @@ namespace MntPlus.WPF
         public string? Frequency => $"Fréquence de lecture du compteur: Chaque {MeterDto?.Frequency} {MeterDto?.FrequencyUnit}";
         public ICommand SaveReadingCommand { get; set; }
         public ICommand DeleteCommand { get; set; }
-        public double ReadingValue { get; set; } 
+        public ICommand EditReadingCommand { get; set; }
+        public double ReadingValue { get; set; }
+
+        public MeterReadingStore MeterReadingStore { get; set; }
+
+        public string ReadingUnite { get; set; } 
         public ReadingMeterViewModel(MeterDto? meterDto , DateTime? nextReading , MeterStore? meterStore)
         {
             CloseCommand = new RelayCommand(async () => await CloseAsync());
             MeterDto = meterDto;
             MeterStore = meterStore;
             NextReading = $"Lecture suivante: {nextReading}";
+            ReadingUnite = MeterDto?.Unit ?? string.Empty;
             SaveReadingCommand = new RelayCommand(async () => await SaveReadingAsync());
             DeleteCommand = new RelayParameterizedCommand(async (p) => await DeleteReadingAsync(p));
+            EditReadingCommand = new RelayParameterizedCommand(async (p) => await EditReadingAsync(p));
             DeleteMeterCommand = new RelayCommand(async () => await DeleteMeterAsync());
             _ = GetReadings();
             EditMeterCommand = new RelayCommand(async () => await EditMeterAsync());
         }
 
+        private async Task EditReadingAsync(object? p)
+        {
+            if (p is MeterReadingDto meterReadingDto)
+            {
+                MeterReadingStore = new MeterReadingStore();
+                MeterReadingStore.MeterReadingUpdated += UpdateReadingValue;
+                var window = new EditReadingWindow();
+                var viewModel = new EditReadingViewModel(MeterReadingStore, meterReadingDto, MeterDto?.Unit!);
+                window.DataContext = viewModel;
+                window.ShowDialog();
+            }
+            await Task.Delay(1);
+        }
+
+        private void UpdateReadingValue(MeterReadingDto meterReadingDto)
+        {
+            var findReading = MeterReadings?.FirstOrDefault(c => c.Id == meterReadingDto.Id);
+            if (findReading != null)
+            {
+                //find index of the reading
+                var index = MeterReadings!.IndexOf(findReading);
+                MeterReadings[index] = meterReadingDto;
+                var lastAddedItem = MeterReadings?.OrderByDescending(c => c.Timestamp).FirstOrDefault();
+                if (lastAddedItem != null && lastAddedItem == meterReadingDto)
+                {
+                    UpdateMeter(new MeterDto(MeterDto!.Id,MeterDto.Name,meterReadingDto.Reading,MeterDto.LastUpdated,MeterDto.Unit,MeterDto.Frequency,
+                        MeterDto.FrequencyUnit,MeterDto.AssetId,MeterDto.AssetName,null)).GetAwaiter().GetResult();
+                }
+               
+            }
+        }
+        private async Task UpdateMeter(MeterDto meterDto)
+        {
+            var result = await AppServices.ServiceManager.MeterService.UpdateMeter(meterDto.Id, 
+                new MeterDtoForCreation(meterDto.Name, meterDto.CurrentReading, meterDto.LastUpdated, meterDto.Unit, meterDto.Frequency, meterDto.AssetId, meterDto.FrequencyUnit), true);
+            if (result.Success)
+            {
+                MeterStore!.UpdateMeter(meterDto);
+            }
+        }
         private async Task EditMeterAsync()
         {
             if (EditAction != null)
@@ -52,12 +99,17 @@ namespace MntPlus.WPF
 
         private async Task DeleteMeterAsync()
         {
+            var Dialog = new ConfirmationWindow("Supprimé Le compteur");
+            Dialog.ShowDialog();
+            if (!Dialog.Confirmed)
+                return;
             await RunCommandAsync(() => SaveIsRunning, async () =>
             {
                 var result = await AppServices.ServiceManager.MeterService.DeleteMeter(MeterDto!.Id,false);
                 if (result.Success)
                 {
                     await IoContainer.NotificationsManager.ShowMessage(new NotificationControlViewModel(NotificationType.Success, "Le compteur a été supprimé avec succès"));
+                    MeterStore!.DeleteMeter(MeterDto);
                     await CloseAsync();
                 }
                 else
@@ -81,17 +133,22 @@ namespace MntPlus.WPF
                 {
                     await RunCommandAsync (() => SaveIsRunning, async () =>
                     {
-                        var result = await AppServices.ServiceManager.MeterReadingService.DeleteUpdateMeterReading(meterReadingDto.Id, false);
+                        var result = await AppServices.ServiceManager.MeterReadingService.DeleteUpdateMeterReading(meterReadingDto.Id, true);
                         if (result.Success)
                         {
                             await IoContainer.NotificationsManager.ShowMessage(new NotificationControlViewModel(NotificationType.Success, "La lecture a été supprimée avec succès"));
                             MeterReadings!.Remove(meterReadingDto);
                             MeterStore!.UpdateMeter(new MeterDto(MeterDto?.Id ?? Guid.Empty, MeterDto?.Name, 0, meterReadingDto.Timestamp, MeterDto!.Unit, MeterDto!.Frequency, MeterDto!.FrequencyUnit, MeterDto!.AssetId,MeterDto.AssetName,null));
                         }
+                        else if(result is ApiBadRequestResponse responseBadRequest)
+                        {
+                            await IoContainer.NotificationsManager.ShowMessage(new NotificationControlViewModel(NotificationType.Error, responseBadRequest.Message));
+                        }
                         else
                         {
                             await IoContainer.NotificationsManager.ShowMessage(new NotificationControlViewModel(NotificationType.Error, "Une erreur s'est produite lors de la suppression de la lecture"));
                         }
+                      
                     });
 
                 }
@@ -99,12 +156,16 @@ namespace MntPlus.WPF
                 {
                     await RunCommandAsync(() => SaveIsRunning, async () =>
                     {
-                        var result = await AppServices.ServiceManager.MeterReadingService.DeleteMeterReading(meterReadingDto.Id, false);
+                        var result = await AppServices.ServiceManager.MeterReadingService.DeleteMeterReading(meterReadingDto.Id, true);
                         if (result.Success)
                         {
                             await IoContainer.NotificationsManager.ShowMessage(new NotificationControlViewModel(NotificationType.Success, "La lecture a été supprimée avec succès"));
                             MeterReadings!.Remove(meterReadingDto);
 
+                        }
+                        else if (result is ApiBadRequestResponse responseBadRequest)
+                        {
+                            await IoContainer.NotificationsManager.ShowMessage(new NotificationControlViewModel(NotificationType.Error, responseBadRequest.Message));
                         }
                         else
                         {
